+++ vclibpy/components/compressors/__init__.py
@@ -2,4 +2,4 @@ from .compressor import Compressor
 from .rotary import RotaryCompressor
 from .ten_coefficient import TenCoefficientCompressor, DataSheetCompressor
 from .constant_effectivness import ConstantEffectivenessCompressor
-from .reciprocating_compresosr import ReciprocatingCompressor
+from .reciprocating_compressor import ReciprocatingCompressor
